Added Files.exists() check for save-file deletion.

+++ foosh/src/main/java/com/vs/foosh/api/services/PersistentDataService.java
@@ -9,6 +9,8 @@ import java.io.ObjectOutputStream;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 
+import org.springframework.boot.ApplicationArguments;
+
 import com.vs.foosh.api.exceptions.misc.CouldNotDeleteCollectionException;
 import com.vs.foosh.api.exceptions.misc.SaveFileNotFoundException;
 import com.vs.foosh.api.exceptions.misc.SavingToFileIOException;
@@ -53,11 +55,13 @@ public class PersistentDataService {
     }
 
     public static void deleteDeviceListSave() {
-        try {
-            Files.copy(ApplicationConfig.getDeviceSavePath(), ApplicationConfig.getDeleteDevicePath(), StandardCopyOption.REPLACE_EXISTING);
-            Files.delete(ApplicationConfig.getDeviceSavePath());
-        } catch (IOException e) {
-            throw new CouldNotDeleteCollectionException();
+        if (Files.exists(ApplicationConfig.getDeviceSavePath())) {
+            try {
+                Files.copy(ApplicationConfig.getDeviceSavePath(), ApplicationConfig.getDeleteDevicePath(), StandardCopyOption.REPLACE_EXISTING);
+                Files.delete(ApplicationConfig.getDeviceSavePath());
+            } catch (IOException e) {
+                throw new CouldNotDeleteCollectionException();
+            }
         }
     }
 
@@ -90,11 +94,13 @@ public class PersistentDataService {
     }
 
     public static void deleteVariableListSave() {
-        try {
-            Files.copy(ApplicationConfig.getVariableSavePath(), ApplicationConfig.getDeleteVariablePath(), StandardCopyOption.REPLACE_EXISTING); //todo
-            Files.delete(ApplicationConfig.getVariableSavePath());
-        } catch (IOException e) {
-            throw new CouldNotDeleteCollectionException();
+        if (Files.exists(ApplicationConfig.getDeleteVariablePath())) {
+            try {
+                Files.copy(ApplicationConfig.getVariableSavePath(), ApplicationConfig.getDeleteVariablePath(), StandardCopyOption.REPLACE_EXISTING); //todo
+                Files.delete(ApplicationConfig.getVariableSavePath());
+            } catch (IOException e) {
+                throw new CouldNotDeleteCollectionException();
+            }
         }
     }
